packages/core/requirements.txt
requests==2.31.0
urllib3==2.2.2
beautifulsoup4==4.12.3
lxml==5.3.0
PyYAML==6.0.2
colorama==0.4.6
playwright==1.47.2
